+++ Scroller/Scroller.js
@@ -16,6 +16,7 @@
  * @exports Scroller
  */
 
+import {setDefaultProps} from '@enact/core/util';
 import {I18nContextDecorator} from '@enact/i18n/I18nDecorator';
 import SpotlightContainerDecorator from '@enact/spotlight/SpotlightContainerDecorator';
 import {ResizeContext} from '@enact/ui/Resizable';
@@ -60,7 +61,7 @@ const scrollerDefaultProps = {
  * @public
  */
 let Scroller = (props) => {
-	const scrollerProps = Object.assign({}, scrollerDefaultProps, props);
+	const scrollerProps = setDefaultProps(props, scrollerDefaultProps);
 
 	// Hooks
 	const {