+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact agate module, newest changes on the top.
 
+## [3.0.0-alpha.2] - 2024-11-21
+
+### Fixed
+
+- `agate/Scroller` and `agate/VirtualList` to have default prop when `undefined` prop is passed
+
 ## [3.0.0-alpha.1] - 2024-08-20
 
 ### Fixed